projects/PoCs/LAIWYER-RAG_Majczyk_Matuszewski/requirements.txt
psycopg2_binary==2.9.10
# PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
sentence_transformers==3.0.1
streamlit==1.32.2
streamlit_extras==0.5.0
